pyproject.toml
[project]
name = "video-summarizer-cli"
version = "0.1.0"
description = "自动下载视频、提取音频、转录并AI总结的命令行工具"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
dependencies = [
    "openai-whisper>=20231117",
    "moviepy>=1.0.3",
    "requests>=2.31.0",
    "yt-dlp>=2023.12.30",
    "pathlib2>=2.3.7; python_version < '3.4'"
]

[project.urls]
Homepage = "https://github.com/your/repo"

[project.optional-dependencies]
test = [
    "pytest"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
video-summarizer = "src.main:main"
audio-summarizer = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta" 